Refactor channel_test
